+++ pso_history/pso_generation_history_1775808976.json
@@ -1,0 +1,152 @@
+[
+  {
+    "generation": 1,
+    "timestamp": "2026-04-10T08:19:02.799317907Z",
+    "best_score": 0.34476493361800165,
+    "weights": {
+      "CpuPercent": -4.026247693846832,
+      "CpuUserSeconds": 1.3004567041257982,
+      "CpuSystemSeconds": -9.866616092484401,
+      "TimezoneDifference": -6.990923907197807,
+      "Distance": 7.0386490643584345,
+      "Radiation": 4.219013860035119,
+      "PowerScore": -7.830914002005277,
+      "TaskCount": -2.054200921398243
+    }
+  },
+  {
+    "generation": 2,
+    "timestamp": "2026-04-10T08:21:46.099698414Z",
+    "best_score": 0.35795755231991133,
+    "weights": {
+      "CpuPercent": -3.2161079517566185,
+      "CpuUserSeconds": 2.3101532915947898,
+      "CpuSystemSeconds": -5.133564118359546,
+      "TimezoneDifference": -6.034068448072677,
+      "Distance": 4.520322701006489,
+      "Radiation": 7.185143677405874,
+      "PowerScore": -3.4553790755237888,
+      "TaskCount": -0.9875534906220071
+    }
+  },
+  {
+    "generation": 3,
+    "timestamp": "2026-04-10T08:24:31.269916972Z",
+    "best_score": 0.35795755231991133,
+    "weights": {
+      "CpuPercent": -3.2161079517566185,
+      "CpuUserSeconds": 2.3101532915947898,
+      "CpuSystemSeconds": -5.133564118359546,
+      "TimezoneDifference": -6.034068448072677,
+      "Distance": 4.520322701006489,
+      "Radiation": 7.185143677405874,
+      "PowerScore": -3.4553790755237888,
+      "TaskCount": -0.9875534906220071
+    }
+  },
+  {
+    "generation": 4,
+    "timestamp": "2026-04-10T08:27:17.034217479Z",
+    "best_score": 0.35795755231991133,
+    "weights": {
+      "CpuPercent": -3.2161079517566185,
+      "CpuUserSeconds": 2.3101532915947898,
+      "CpuSystemSeconds": -5.133564118359546,
+      "TimezoneDifference": -6.034068448072677,
+      "Distance": 4.520322701006489,
+      "Radiation": 7.185143677405874,
+      "PowerScore": -3.4553790755237888,
+      "TaskCount": -0.9875534906220071
+    }
+  },
+  {
+    "generation": 5,
+    "timestamp": "2026-04-10T08:30:02.527467982Z",
+    "best_score": 0.35795755231991133,
+    "weights": {
+      "CpuPercent": -3.2161079517566185,
+      "CpuUserSeconds": 2.3101532915947898,
+      "CpuSystemSeconds": -5.133564118359546,
+      "TimezoneDifference": -6.034068448072677,
+      "Distance": 4.520322701006489,
+      "Radiation": 7.185143677405874,
+      "PowerScore": -3.4553790755237888,
+      "TaskCount": -0.9875534906220071
+    }
+  },
+  {
+    "generation": 6,
+    "timestamp": "2026-04-10T08:33:14.018563111Z",
+    "best_score": 0.36363632381653826,
+    "weights": {
+      "CpuPercent": -2.6064898611182965,
+      "CpuUserSeconds": 3.988287998115447,
+      "CpuSystemSeconds": -4.791145372327378,
+      "TimezoneDifference": -5.902888194894891,
+      "Distance": 4.197657735931452,
+      "Radiation": 7.6140866644517,
+      "PowerScore": -2.9776647265451635,
+      "TaskCount": -1.1706840745970186
+    }
+  },
+  {
+    "generation": 7,
+    "timestamp": "2026-04-10T08:36:02.158657628Z",
+    "best_score": 0.36363632381653826,
+    "weights": {
+      "CpuPercent": -2.6064898611182965,
+      "CpuUserSeconds": 3.988287998115447,
+      "CpuSystemSeconds": -4.791145372327378,
+      "TimezoneDifference": -5.902888194894891,
+      "Distance": 4.197657735931452,
+      "Radiation": 7.6140866644517,
+      "PowerScore": -2.9776647265451635,
+      "TaskCount": -1.1706840745970186
+    }
+  },
+  {
+    "generation": 8,
+    "timestamp": "2026-04-10T08:38:53.681428417Z",
+    "best_score": 0.36363632381653826,
+    "weights": {
+      "CpuPercent": -2.6064898611182965,
+      "CpuUserSeconds": 3.988287998115447,
+      "CpuSystemSeconds": -4.791145372327378,
+      "TimezoneDifference": -5.902888194894891,
+      "Distance": 4.197657735931452,
+      "Radiation": 7.6140866644517,
+      "PowerScore": -2.9776647265451635,
+      "TaskCount": -1.1706840745970186
+    }
+  },
+  {
+    "generation": 9,
+    "timestamp": "2026-04-10T08:41:42.795434339Z",
+    "best_score": 0.36363632381653826,
+    "weights": {
+      "CpuPercent": -2.6064898611182965,
+      "CpuUserSeconds": 3.988287998115447,
+      "CpuSystemSeconds": -4.791145372327378,
+      "TimezoneDifference": -5.902888194894891,
+      "Distance": 4.197657735931452,
+      "Radiation": 7.6140866644517,
+      "PowerScore": -2.9776647265451635,
+      "TaskCount": -1.1706840745970186
+    }
+  },
+  {
+    "generation": 10,
+    "timestamp": "2026-04-10T08:44:34.634686796Z",
+    "best_score": 0.36363632381653826,
+    "weights": {
+      "CpuPercent": -2.6064898611182965,
+      "CpuUserSeconds": 3.988287998115447,
+      "CpuSystemSeconds": -4.791145372327378,
+      "TimezoneDifference": -5.902888194894891,
+      "Distance": 4.197657735931452,
+      "Radiation": 7.6140866644517,
+      "PowerScore": -2.9776647265451635,
+      "TaskCount": -1.1706840745970186
+    }
+  }
+]
